+++ dev/Makefile
@@ -6,39 +6,39 @@
 #    By: akurochk <akurochk@student.42.fr>          +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2024/04/11 11:00:13 by akurochk          #+#    #+#              #
-#    Updated: 2024/04/25 10:42:13 by akurochk         ###   ########.fr        #
+#    Updated: 2024/04/25 14:11:32 by akurochk         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
 NAME	= minishell
 
 SRCS	= main.c \
-	builtins_call.c \
-	command_call.c \
-	common.c \
-	env.c \
-	error_handler.c \
-	executor.c \
-	ft_itoa.c \
-	ft_split.c \
-	ft_str.c \
-	ft_substr.c \
-	group.c \
-	heredoc.c \
-	lexer.c \
-	list.c \
-	parser_star.c \
-	parser.c \
-	pipes.c \
-	signal.c \
-	subshell.c
+	builtins/builtins_call.c \
+	builtins/command_call.c \
+	parser/common.c \
+	utils/env.c \
+	utils/error_handler.c \
+	parser/executor.c \
+	utils/ft_itoa.c \
+	utils/ft_split.c \
+	utils/ft_str.c \
+	utils/ft_substr.c \
+	parser/group.c \
+	parser/heredoc.c \
+	lexer/lexer.c \
+	list/list.c \
+	parser/parser_star.c \
+	parser/parser.c \
+	parser/pipes.c \
+	utils/signal.c \
+	parser/subshell.c
 #	test_cmp.c
 #	test_lexer.c
 
 OBJS	= $(SRCS:.c=.o)
 
 CC		= cc
-CFLAGS	= -g3 -Wall -Werror -Wextra
+CFLAGS	= -g3 -Wall -Werror -Wextra -I ./inc
 
 all: $(NAME)
