+++ dev/parser/heredoc.c
@@ -6,11 +6,11 @@
 /*   By: akurochk <akurochk@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/23 13:58:50 by akurochk          #+#    #+#             */
-/*   Updated: 2024/04/25 12:37:50 by akurochk         ###   ########.fr       */
+/*   Updated: 2024/04/25 14:06:53 by akurochk         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "minishell.h"
+#include "../inc/minishell.h"
 
 char	*ft_str_concat(const char *s1, char c, const char *s2, char use_c)
 {
