+++ dev/parser/pipes.c
@@ -6,11 +6,11 @@
 /*   By: akurochk <akurochk@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/23 16:46:02 by akurochk          #+#    #+#             */
-/*   Updated: 2024/04/25 12:06:53 by akurochk         ###   ########.fr       */
+/*   Updated: 2024/04/25 14:06:57 by akurochk         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "minishell.h"
+#include "../inc/minishell.h"
 
 int	get_fd_in(t_cmd_info *cmd_info)
 {
