+++ dev/tests/test_cmp.c
