+++ dev/tests/test_lexer.c
@@ -6,11 +6,11 @@
 /*   By: akurochk <akurochk@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/17 18:55:29 by akurochk          #+#    #+#             */
-/*   Updated: 2024/04/18 14:39:58 by akurochk         ###   ########.fr       */
+/*   Updated: 2024/04/25 14:06:59 by akurochk         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "minishell.h"
+#include "../inc/minishell.h"
 
 void	test_print_tokens(t_list *toks)
 {
