+++ dev/utils/error_handler.c
@@ -6,11 +6,11 @@
 /*   By: akurochk <akurochk@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/16 20:13:17 by akurochk          #+#    #+#             */
-/*   Updated: 2024/04/25 12:09:10 by akurochk         ###   ########.fr       */
+/*   Updated: 2024/04/25 14:07:05 by akurochk         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "minishell.h"
+#include "../inc/minishell.h"
 
 /* 
 ret		- return val of function.
