+++ dev/utils/signal.c
@@ -1,38 +1,39 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_str.c                                           :+:      :+:    :+:   */
+/*   signal.c                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: akurochk <akurochk@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/05/03 15:51:38 by akurochk          #+#    #+#             */
-/*   Updated: 2024/04/25 13:54:14 by akurochk         ###   ########.fr       */
+/*   Created: 2024/04/16 17:25:30 by akurochk          #+#    #+#             */
+/*   Updated: 2024/04/25 14:07:09 by akurochk         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "minishell.h"
+#include "../inc/minishell.h"
 
-int	ft_strlen(const char *s)
+void	handler_signal(int signum)
 {
-	int	len;
-
-	len = 0;
-	while (s[len])
-		len++;
-	return (len);
+	if (signum == SIGINT)
+	{
+		write(1, "\n", 1);
+		rl_on_new_line();
+		rl_replace_line("", 1);
+		rl_redisplay();
+	}
 }
 
-char	*ft_strchr(const char *s, int c)
+void	handler_heredoc(int signum)
 {
-	if (c < 0)
-		return (NULL);
-	while (*s)
-	{
-		if (*s == (unsigned char)c)
-			return ((char *)s);
-		s++;
-	}
-	if (*s == (unsigned char)c)
-		return ((char *)s);
-	return (NULL);
+	(void)signum;
+	write(2, "\n", 1);
+	exit (130);
+}
+
+void	handler_executor(int signum)
+{
+	if (signum == SIGINT)
+		write(2, "\n", 1);
+	else if (signum == SIGQUIT)
+		write(2, "Exit: handler_executor\n", 5);
 }
